feat: add Responses API support to AzureLLM via shared OpenAICompatibleLLM base class
